start to implement main part

+++ include/check_files.h
@@ -8,6 +8,6 @@
 #define DEFAULT_MAX_ARGS 4
 #define DEFAULT_MAX_FUNC 10
 
-int check_file(FILE *file, size_t max_size , size_t max_args, size_t max_func);
+size_t check_file(FILE *file, size_t max_size , size_t max_args, size_t max_func);
 
 #endif /* ! CHECK_FILES_H */
